Adapt `_filter_image_list`/`match_filter` to new filters objects
