+++ fractal_server/app/runner/v2/filters.py
@@ -1,0 +1,9 @@
+from typing import Any
+
+from pydantic import BaseModel
+from pydantic import Field
+
+
+class Filters(BaseModel):
+    attributes: dict[str, Any] = Field(default_factory=dict)
+    types: dict[str, bool] = Field(default_factory=dict)
